+++ R/chk-datetime.R
@@ -20,16 +20,20 @@ chk_date_time <- function(x, x_name = NULL) {
     return(invisible(x))
   }
   if (is.null(x_name)) x_name <- deparse_backtick_chk(substitute(x))
-  abort_chk(x_name, " must be a date time (non-missing POSIXct scalar)", x = x)}
+  abort_chk(x_name, " must be a date time (non-missing POSIXct scalar)", x = x)
+}
 
 #' @describeIn chk_date_time Check Date Time (Deprecated)
 #'
+#' \lifecycle{deprecated}
+#'
 #' @export
 chk_datetime <- function(x, x_name = NULL) {
-  deprecate_soft("0.4.1",
-                 what = "chk_datetime()",
-                 with = "chk_date_time()",
-                 id = "chk_datetime")
+  deprecate_warn("0.4.1",
+    what = "chk::chk_datetime()",
+    with = "chk::chk_date_time()",
+    id = "chk_datetime"
+  )
   if (vld_date_time(x)) {
     return(invisible(x))
   }
@@ -53,11 +57,14 @@ vld_date_time <- function(x) {
 
 #' @describeIn chk_date_time Validate Date Time (Deprecated)
 #'
+#' \lifecycle{deprecated}
+#'
 #' @export
 vld_datetime <- function(x) {
-  deprecate_soft("0.4.1",
-                 what = "vld_datetime()",
-                 with = "vld_date_time()",
-                 id = "chk_datetime")
+  deprecate_warn("0.4.1",
+    what = "chk::vld_datetime()",
+    with = "chk::vld_date_time()",
+    id = "chk_datetime"
+  )
   vld_date_time(x)
 }